Section safe subgraph generation tool
Subgraph Uncrashable is a codegen tool that makes loading and interacting with entities atomic, safe and consistent. Wave 6 successful Grantee contribution (Safe Subgraph Code Generator - Float - $50,000), as part of project closure, it was agreed to push a summary of the tool to the thegraph/com docs. codegen section of the docs seemed most applicable.